+++ R/dissimilarity.R
@@ -79,8 +79,8 @@
 #' @examples
 #' comat <- matrix(sample(0:1000, size = 50, replace = TRUE,
 #' prob = 1 / 1:1001), 5, 10)
-#' rownames(comat) <- paste0("Site", 1:5)
-#' colnames(comat) <- paste0("Species", 1:10)
+#' rownames(comat) <- paste0("s", 1:5)
+#' colnames(comat) <- paste0("sp", 1:10)
 #'
 #' dissim <- dissimilarity(comat,
 #' metric = c("abc", "ABC", "Simpson", "Brayturn"))