+++ R/similarity_dissimilarity_conversion.R
@@ -50,8 +50,8 @@
 #' @examples
 #' comat <- matrix(sample(0:1000, size = 50, replace = TRUE,
 #' prob = 1 / 1:1001), 5, 10)
-#' rownames(comat) <- paste0("Site", 1:5)
-#' colnames(comat) <- paste0("Species", 1:10)
+#' rownames(comat) <- paste0("s", 1:5)
+#' colnames(comat) <- paste0("sp", 1:10)
 #'
 #' simil <- similarity(comat, metric = "all")
 #' simil
@@ -162,8 +162,8 @@ similarity_to_dissimilarity <- function(similarity, include_formula = TRUE){
 #' @examples
 #' comat <- matrix(sample(0:1000, size = 50, replace = TRUE,
 #' prob = 1 / 1:1001), 5, 10)
-#' rownames(comat) <- paste0("Site", 1:5)
-#' colnames(comat) <- paste0("Species", 1:10)
+#' rownames(comat) <- paste0("s", 1:5)
+#' colnames(comat) <- paste0("sp", 1:10)
 #'
 #' dissimil <- dissimilarity(comat, metric = "all")
 #' dissimil